moving propagation input to separated file
